+++ MedataCore/Tests/HarnessCLITests/AccuracyHarnessTests.swift
@@ -1,3 +1,4 @@
+#if HARNESS_ENABLED
 import XCTest
 import Foods
 import PortableContracts
@@ -141,3 +142,4 @@ final class AccuracyHarnessTests: XCTestCase {
         )
     }
 }
+#endif