+++ MedataCore/Tests/HarnessCLITests/BetaCalibratorTests.swift
@@ -1,3 +1,4 @@
+#if HARNESS_ENABLED
 import XCTest
 import Foods
 import PortableContracts
@@ -155,3 +156,4 @@ final class BetaCalibratorTests: XCTestCase {
         }
     }
 }
+#endif